# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Vaulta API
  version: 1.0.0
  description: Secure file metadata, multipart upload authorization, and sharing API.
servers:
  - url: http://localhost:4000
tags:
  - name: Authentication
  - name: Files
  - name: Public
paths:
  /api/auth/register:
    post:
      tags: [Authentication]
      summary: Create an account and authenticated session
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Credentials' }
      responses:
        '201':
          description: Account created
          headers:
            Set-Cookie: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /api/auth/login:
    post:
      tags: [Authentication]
      summary: Create an authenticated session
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Credentials' }
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /api/auth/me:
    get:
      tags: [Authentication]
      summary: Read the current session user
      security: [{ sessionCookie: [] }]
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/auth/logout:
    post:
      tags: [Authentication]
      summary: Revoke the current session
      security: [{ sessionCookie: [], csrfHeader: [] }]
      responses:
        '204': { description: Session revoked }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/files:
    get:
      tags: [Files]
      summary: List the current user's files
      security: [{ sessionCookie: [] }]
      parameters:
        - in: query
          name: cursor
          schema: { type: string, format: uuid }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
      responses:
        '200':
          description: Cursor page
          content:
            application/json:
              schema:
                type: object
                required: [items, nextCursor]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/File' }
                  nextCursor: { type: [string, 'null'], format: uuid }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/files/uploads:
    post:
      tags: [Files]
      summary: Initialize a private multipart upload
      security: [{ sessionCookie: [], csrfHeader: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [originalName, mimeType, sizeBytes]
              properties:
                originalName: { type: string, maxLength: 255 }
                mimeType:
                  type: string
                  enum: [application/pdf, image/png, image/jpeg, text/plain, application/zip]
                sizeBytes: { type: integer, minimum: 1, maximum: 262144000 }
      responses:
        '201':
          description: Upload initialized
          content:
            application/json:
              schema:
                type: object
                required: [file, partSizeBytes, partCount]
                properties:
                  file: { $ref: '#/components/schemas/File' }
                  partSizeBytes: { type: integer }
                  partCount: { type: integer }
        '413': { $ref: '#/components/responses/ValidationError' }
        '415': { $ref: '#/components/responses/ValidationError' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /api/files/{fileId}/parts:
    post:
      tags: [Files]
      summary: Create short-lived upload URLs for selected parts
      security: [{ sessionCookie: [], csrfHeader: [] }]
      parameters:
        - $ref: '#/components/parameters/FileId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [partNumbers]
              properties:
                partNumbers:
                  type: array
                  minItems: 1
                  maxItems: 1000
                  uniqueItems: true
                  items: { type: integer, minimum: 1 }
      responses:
        '200':
          description: Signed parts
          content:
            application/json:
              schema:
                type: object
                properties:
                  parts:
                    type: array
                    items:
                      type: object
                      required: [partNumber, url]
                      properties:
                        partNumber: { type: integer }
                        url: { type: string, format: uri }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/files/{fileId}/complete:
    post:
      tags: [Files]
      summary: Assemble, validate, and publish the upload
      security: [{ sessionCookie: [], csrfHeader: [] }]
      parameters:
        - $ref: '#/components/parameters/FileId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [parts]
              properties:
                parts:
                  type: array
                  minItems: 1
                  maxItems: 1000
                  items:
                    type: object
                    additionalProperties: false
                    required: [partNumber, etag]
                    properties:
                      partNumber: { type: integer, minimum: 1 }
                      etag: { type: string, maxLength: 128 }
      responses:
        '200':
          description: Completed file
          content:
            application/json:
              schema:
                type: object
                properties:
                  file: { $ref: '#/components/schemas/File' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /api/files/{fileId}:
    patch:
      tags: [Files]
      summary: Change file visibility
      security: [{ sessionCookie: [], csrfHeader: [] }]
      parameters:
        - $ref: '#/components/parameters/FileId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [visibility]
              properties:
                visibility: { type: string, enum: [PRIVATE, PUBLIC] }
      responses:
        '200':
          description: Updated file
          content:
            application/json:
              schema:
                type: object
                properties:
                  file: { $ref: '#/components/schemas/File' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
    delete:
      tags: [Files]
      summary: Abort an upload or delete a file
      security: [{ sessionCookie: [], csrfHeader: [] }]
      parameters:
        - $ref: '#/components/parameters/FileId'
      responses:
        '204': { description: Deleted }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/files/{fileId}/download:
    get:
      tags: [Files]
      summary: Create a short-lived owner download URL
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: '#/components/parameters/FileId'
      responses:
        '200':
          description: Download authorization
          content:
            application/json:
              schema:
                type: object
                required: [url, expiresIn]
                properties:
                  url: { type: string, format: uri }
                  expiresIn: { type: integer, example: 300 }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/public/{shareToken}:
    get:
      tags: [Public]
      summary: Redirect a valid public token to a short-lived download URL
      parameters:
        - in: path
          name: shareToken
          required: true
          schema: { type: string, pattern: '^[A-Za-z0-9_-]{43}$' }
      responses:
        '302':
          description: Temporary redirect
          headers:
            Location: { schema: { type: string, format: uri } }
        '404': { $ref: '#/components/responses/NotFound' }
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: secure_store_session
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
  parameters:
    FileId:
      in: path
      name: fileId
      required: true
      schema: { type: string, format: uuid }
  schemas:
    Credentials:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        email: { type: string, format: email, maxLength: 254 }
        password: { type: string, minLength: 12, maxLength: 128 }
    User:
      type: object
      required: [id, email, createdAt]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        createdAt: { type: string, format: date-time }
    AuthResponse:
      type: object
      required: [user]
      properties:
        user: { $ref: '#/components/schemas/User' }
    File:
      type: object
      required: [id, originalName, mimeType, sizeBytes, status, visibility, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        originalName: { type: string }
        mimeType: { type: string }
        sizeBytes: { type: integer }
        status: { type: string, enum: [UPLOADING, READY, REJECTED] }
        visibility: { type: string, enum: [PRIVATE, PUBLIC] }
        rejectionReason: { type: [string, 'null'] }
        shareUrl: { type: [string, 'null'], format: uri }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code: { type: string }
            message: { type: string }
            requestId: { type: string }
            details: {}
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: CSRF token missing or invalid
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource not found or not owned
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: State conflict
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }